[SWM-324] Refactor : remove existing member service
